Cambiada la ruta de la vista & creado componente PagesSlides
