+++ src/components/index.js
@@ -1,1 +1,2 @@
 export { default as Form } from './Form'
+export { default as PageSlides } from './PagesSlides'
